bfd/ 	* bfd-in.h (bfd_elf_bfd_from_remote_memory): Make LEN argument 	of target_read_memory as size_t. 	* bfd-in2.h: Regenerate. 	* elf-bfd.h (elf_backend_bfd_from_remote_memory): Make LEN 	argument of target_read_memory as size_t. 	(_bfd_elf32_bfd_from_remote_memory): Likewise. 	(_bfd_elf64_bfd_from_remote_memory): Likewise. 	* elf.c (bfd_elf_bfd_from_remote_memory): Likewise. 	* elfcode.h (NAME(_bfd_elf,bfd_from_remote_memory)): Likewise.
gdb/
	* target.c (target_read_memory): Make LEN argument as size_t.
	* target.h (target_read_memory): Likewise.

+++ bfd/bfd-in.h
@@ -698,7 +698,7 @@ extern int bfd_get_elf_phdrs
    the remote memory.  */
 extern bfd *bfd_elf_bfd_from_remote_memory
   (bfd *templ, bfd_vma ehdr_vma, bfd_vma *loadbasep,
-   int (*target_read_memory) (bfd_vma vma, bfd_byte *myaddr, int len));
+   int (*target_read_memory) (bfd_vma vma, bfd_byte *myaddr, size_t len));
 
 extern struct bfd_section *_bfd_elf_tls_setup
   (bfd *, struct bfd_link_info *);
